@@ -1,6 +1,7 @@
 {
-  "name": "nextra-docs-template",
-  "version": "0.0.1",
+  "name": "nextra-v3-docs-template",
+  "version": "0.0.2",
+  "type": "module",
   "description": "Nextra docs template",
   "scripts": {
     "dev": "next dev",
@@ -18,14 +19,14 @@
   },
   "homepage": "https://github.com/shuding/nextra-docs-template#readme",
   "dependencies": {
-    "next": "^13.0.6",
-    "nextra": "latest",
-    "nextra-theme-docs": "latest",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "next": "^15.1.3",
+    "nextra": "^3.3.1",
+    "nextra-theme-docs": "^3.3.1",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0"
   },
   "devDependencies": {
-    "@types/node": "18.11.10",
-    "typescript": "^4.9.3"
+    "@types/node": "22.10.2",
+    "typescript": "^5.7.2"
   }
 }
